[RENAME] core->common 패키지로 변경
