+++ src/main/java/com/programmers/smrtstore/domain/order/infrastructure/OrderJpaRepository.java
@@ -8,24 +8,60 @@ import org.springframework.data.jpa.repository.Query;
 
 public interface OrderJpaRepository extends JpaRepository<Order, Long>, OrderRepositoryCustom {
 
-    // TODO: user 가 탈퇴한 경우 어떻게 처리할 것인가?
+    // TODO: user 가 탈퇴한 경우 해당 메소드 어떻게 처리할 것인가?
+    /**
+     * 단일 Order를 조회합니다. 이 메소드는 삭제 된 Order 엔티티는 제외 하고 반환합니다. 이 메소드는 'orderSheet'와 'orderSheet 의 user'
+     * 필드를 Eager로 로딩합니다.
+     *
+     * @param orderId 조회하고자 하는 Order의 ID입니다.
+     * @return Optional<Order> - 해당 ID를 가진 Order 객체. 만약 Order가 존재하지 않거나 'deletedAt'이 null이 아닐 경우, 비어
+     * 있는 Optional을 반환합니다.
+     */
     @EntityGraph(attributePaths = {"orderSheet", "orderSheet.user"})
     @Query("SELECT o FROM Order o WHERE o.id = :orderId AND o.deletedAt IS NULL")
     Optional<Order> findByIdWithOrderSheetAndUser(Long orderId);
 
+    /**
+     * 단일 Order를 조회합니다. 이 메소드는 삭제 된 Order 엔티티는 제외 하고 반환합니다. 이 메소드는 'orderSheet' 필드를 Eager로 로딩합니다.
+     *
+     * @param orderId 조회하고자 하는 Order의 ID입니다.
+     * @return Optional<Order> - 해당 ID를 가진 Order 객체. 만약 Order가 존재하지 않거나 'deletedAt'이 null이 아닐 경우, 비어
+     * 있는 Optional을 반환합니다.
+     */
     @EntityGraph(attributePaths = {"orderSheet"})
     @Query("SELECT o FROM Order o WHERE o.id = :orderId AND o.deletedAt IS NULL")
     Optional<Order> findByIdWithOrderSheet(Long orderId);
 
+    /**
+     * 단일 Order를 조회합니다. 이 메소드는 삭제 된 Order 엔티티는 제외 하고 반환합니다.
+     *
+     * @param orderId 조회하고자 하는 Order의 ID입니다.
+     * @return Optional<Order> - 해당 ID를 가진 Order 객체. 만약 Order가 존재하지 않거나 'deletedAt'이 null이 아닐 경우, 비어
+     * 있는 Optional을 반환합니다.
+     */
     @Override
     @Query("SELECT o FROM Order o WHERE o.id = :orderId AND o.deletedAt IS NULL")
     Optional<Order> findById(Long orderId);
 
+    /**
+     * 단일 Order를 조회합니다. 이 메소드는 삭제 된 Order 엔티티도 포함하여 결과를 반환합니다. 이 메소드는 'orderSheet' 필드를 Eager로
+     * 로딩합니다.
+     *
+     * @param orderId 조회하고자 하는 Order의 ID입니다.
+     * @return Optional<Order> - 해당 ID를 가진 Order 객체. 만약 Order가 존재하지 않으면, 비어 있는 Optional을 반환합니다.
+     */
     @EntityGraph(attributePaths = {"orderSheet"})
     @Query("SELECT o FROM Order o WHERE o.id = :orderId")
     Optional<Order> findByIdWithOrderSheetIncludeDeleted(Long orderId);
 
-    @Query("SELECT o FROM Order o WHERE o.id = :orderIdL")
-    Optional<Order> findByIdIncludeDeleted(Long orderIdL);
+    /**
+     * 단일 Order를 조회합니다. 이 메소드는 삭제 된 Order 엔티티도 포함하여 결과를 반환합니다.
+     *
+     * @param orderId 조회하고자 하는 Order의 ID입니다.
+     * @return Optional<Order> - 해당 ID를 가진 Order 객체. 만약 Order가 존재하지 않으면, 비어
+     * 있는 Optional을 반환합니다.
+     */
+    @Query("SELECT o FROM Order o WHERE o.id = :orderId")
+    Optional<Order> findByIdIncludeDeleted(Long orderId);
 
 }
